// client/ts/src/wrapper/types/MarketInfo.ts
/**
 * This code was GENERATED using the solita package.
 * Please DO NOT EDIT THIS FILE, instead rerun solita to update it or write a wrapper to add functionality.
 *
 * See: https://github.com/metaplex-foundation/solita
 */

import * as web3 from '@solana/web3.js';
import * as beet from '@metaplex-foundation/beet';
import * as beetSolana from '@metaplex-foundation/beet-solana';
export type MarketInfo = {
  market: web3.PublicKey;
  ordersRootIndex: number;
  traderIndex: number;
  baseBalance: beet.bignum;
  quoteBalance: beet.bignum;
  quoteVolume: beet.bignum;
  lastUpdatedSlot: number;
  padding: number[] /* size: 3 */;
};

/**
 * @category userTypes
 * @category generated
 */
export const marketInfoBeet = new beet.BeetArgsStruct<MarketInfo>(
  [
    ['market', beetSolana.publicKey],
    ['ordersRootIndex', beet.u32],
    ['traderIndex', beet.u32],
    ['baseBalance', beet.u64],
    ['quoteBalance', beet.u64],
    ['quoteVolume', beet.u64],
    ['lastUpdatedSlot', beet.u32],
    ['padding', beet.uniformFixedSizeArray(beet.u32, 3)],
  ],
  'MarketInfo',
);